requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytesseract
python-multipart
Pillow
google-genai
pydantic